Move category items to apihelper
